+++ weaver/wps_restapi/utils.py
@@ -8,7 +8,7 @@ import requests
 import logging
 if TYPE_CHECKING:
     from pyramid.request import Request
-    from weaver.typedefs import AnySettingsContainer
+    from weaver.typedefs import AnyAnySettingsContainer
 
 LOGGER = logging.getLogger("weaver")
 
@@ -25,14 +25,14 @@ OUTPUT_FORMATS = {
 
 
 def wps_restapi_base_path(container):
-    # type: (AnySettingsContainer) -> AnyStr
+    # type: (AnyAnySettingsContainer) -> AnyStr
     settings = get_settings(container)
     restapi_path = settings.get("weaver.wps_restapi_path", "").rstrip('/').strip()
     return restapi_path
 
 
 def get_wps_restapi_base_url(container):
-    # type: (AnySettingsContainer) -> AnyStr
+    # type: (AnyAnySettingsContainer) -> AnyStr
     settings = get_settings(container)
     weaver_url = get_weaver_url(settings)
     restapi_path = wps_restapi_base_path(settings)